Added popToTop() in the onReplace method
Added `this.refs.nav.popToTop()` to the onReplace method of `ExRouter.js` so that the components **unmount** when replacing the current route with another route.  I was having an issue where one of my child routes that requires an event listener wasn't having it's `componentWillUnmount` called when replacing the route.  This change is working for my use case.

I'm new to this package so there may be a better way to solve this.