fix(NavBar): don't render custom BackButton if root nav state (#998)
Avoid rendering a custom `BackButton` if the default implementation
should avoid rendering, such as when the current scene is the root
scene. Since main `state` isn't being passed to the custom component
we have no way (that I'm aware of) of determining whether it should
actually render a back button.